Fix isinstance on curation.models.ContentType
Makes it so that isinstance(curation.models.ContentType, obj)
returns True if obj is an instance of the concrete ContentType model.

If this is not done, a ValueError is thrown if the user passes a concrete
ContentType instance to a field with rel.to=curation.models.ContentType (e.g.
curation.fields.ContentTypeSourceField)