Not all models have DoesNotExist?
Ran into this trying to do Wire curation. It could be a bug on the Wire's end though. My understanding is that everything inheriting from model ought to have DoesNotExist, but I think that the skipping done in the metaclass might make this not the case for things inheriting from CuratedItem.