dj22-compat: Prevent django checks raising fields.E005 warning
The fields.E005 check ('choices' must be an iterable (actual value,
human readable name) tuples.) fails for django-curation, because it uses
a slightly different datastructure so as to pass more data to the
choices widget. This choices object is programmatically generated, so
it is sufficient to disable this check, the purpose of which is to catch
developer errors in the specification of the choices kwarg.